# swagger.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "FinVeda API",
                 "version":  "v1"
             },
    "paths":  {
...
                  "/api/v1/Product":  {
                                          "post":  {
                                                       "tags":  [
                                                                    "Product"
                                                                ],
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  "@{schema=}",
                                                                                           "text/json":  "@{schema=}",
                                                                                           "application/*+json":  "@{schema=}"
                                                                                       }
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK"
                                                                                 }
                                                                     }
                                                   }
                                      },
                  "/api/v1/Product/{id}":  {
                                               "delete":  {
                                                              "tags":  [
                                                                           "Product"
                                                                       ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  "@{type=string; format=uuid}"
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK"
                                                                                        }
                                                                            }
                                                          }
                                           },
...
                  "/api/v1/Report/par":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "Report"
                                                                  ],
                                                         "parameters":  [
                                                                            {
                                                                                "name":  "start",
                                                                                "in":  "query",
                                                                                "schema":  "@{type=string; format=date-time}"
                                                                            },
                                                                            {
                                                                                "name":  "end",
                                                                                "in":  "query",
                                                                                "schema":  "@{type=string; format=date-time}"
                                                                            }
                                                                        ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK"
                                                                                   }
                                                                       }
                                                     }
                                         },
                  "/api/v1/Report/efficiency":  {
                                                    "get":  {
                                                                "tags":  [
                                                                             "Report"
                                                                         ],
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "start",
                                                                                       "in":  "query",
                                                                                       "schema":  "@{type=string; format=date-time}"
                                                                                   },
                                                                                   {
                                                                                       "name":  "end",
                                                                                       "in":  "query",
                                                                                       "schema":  "@{type=string; format=date-time}"
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK"
                                                                                          }
                                                                              }
                                                            }
                                                },
...
              },
    "components":  {
...
                       "securitySchemes":  {
                                               "Bearer":  {
                                                              "type":  "http",
                                                              "description":  "Enter \u0027Bearer\u0027 [space] and then your valid token.",
                                                              "scheme":  "Bearer",
                                                              "bearerFormat":  "JWT"
                                                          }
                                           }
                   },
    "security":  [
                     {
                         "Bearer":  [

                                    ]
                     }
                 ]
}